Main game loop mechanic completed
The game loop manager is editable in the scene. At first it will initialize every variables.

By adjusting the Time Limit the time of each loop can be lengthen or shorten, default please set to 300s and set the decrement of Time limit to 30s.

Total progress is a variable to check How far the player still need to go to break the time loop and finally win, default is 0 since it is for display.

Progress needed is a variable to check how many tasks has the player done, it will start at 3 for easy gameplay and increase by iteration until it is 6, which is also the estimated amount of puzzles.

Progress in this iteration is a variable to show the current progress in this iteration, default is 0 since it is for display.

Total iterations decide how many iteration the player needs to go through, default please set to 6 for longer gameplay.

Max Attempts represents how many times player are allowed to try for a fail attempt, by default it will not connect to any puzzle reset logic for easier gameplay experience. Default please set it to 5 or 6 is okay.

Current Attempt represents the current attempt, once the player fail to reach the progress needed, the situation is considered attempt failed and the Current Attempt increase by 1. Until it is equal or larger than Max Attempts, then it is considered failed, default is 0 since it is for display.

Time toString is for easy parsing value to UI for the countdown timer, use it or delete it if a better method is used.

SecondPuzzle_Array is an attempt to resetting puzzle after the player proceed to the next iteration. According to the nature of different puzzles, they may or may not need to be reset. For example First Puzzle will create a new widget every time and hence it doesn't required resetting.

-------------------------------------------------------------------------------

Every game loop will reset the iteration after the timer reached 0. Logic will be checked as well as the winning and losing mechanic.

Reset all puzzles if player proceeds to next level (Time limit decreased)

Then reset player position and rotation.

Then losing and winning condition check

Losing Condition Checking:
Current Attempt >= Max Attempts
if the condition didn't match, do winning condition checking

Winning Condition checking:
Total Progress >= Total Iterations